refactor(gsd): extract auto-worktree entry path module
